LPD-30906 Rename undoMoveItem by undoMoveItems
